Add global tick generator Rework components to use global tick generator
