+++ project.godot
@@ -22,6 +22,7 @@ config/icon="res://icon.svg"
 [autoload]
 
 SessionUserSettings="*res://src/global/SessionUserSettings.cs"
+TickGenerator="*res://src/global/TickGenerator.cs"
 
 [display]
