+++ src/entities/components/EngineComponent.cs
@@ -32,6 +32,7 @@ public partial class EngineComponent : ConsumerComponent
     // Called when the node enters the scene tree for the first time.
     public override void _Ready()
     {
+        base._Ready();
         currentHeat = 0;
     }
