+++ src/entities/components/PowerPlantComponent.cs
@@ -62,6 +62,7 @@ public partial class PowerPlantComponent : ConsumerComponent
     // Called when the node enters the scene tree for the first time.
     public override void _Ready()
     {
+        base._Ready();
         ComponentsChanged();
     }
